# perf(trie-router): fine tuning, 9~10% faster