I did some minor tweaking of the TrieRouter and made it 9-10% faster.

```
hono - trie-router(default) x 434,571 ops/sec ±4.82% (81 runs sampled)
hono - [Optimized] trie-router(default) x 465,051 ops/sec ±3.45% (87 runs sampled)
hono - regexp-router x 528,354 ops/sec ±4.34% (84 runs sampled)
itty-router x 201,821 ops/sec ±4.02% (88 runs sampled)
sunder x 320,189 ops/sec ±2.17% (88 runs sampled)
worktop x 191,326 ops/sec ±2.88% (83 runs sampled)
Fastest is hono - regexp-router
✨  Done in 44.36s.
```